requirements.txt
# Backend
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
pydantic-settings==2.4.0
aiohttp==3.10.5

# Bot
python-telegram-bot==21.5
python-dotenv==1.0.1
